refactor: extract updater functions
